# app/backend/openapi/openapi.v1.yaml
openapi: 3.0.3
info:
  title: AeroGenIoT API
  version: 1.0.0
  description: API base para AeroGenIoT (monitor de cultivos)
servers:
  - url: /api/v1
...
paths:
...
  /thresholds:
    get:
      tags: [thresholds]
      summary: Obtiene umbrales por granja/plot
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: query
          required: true
          schema:
            type: string
        - name: plotId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThresholdProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [thresholds]
      summary: Actualiza umbrales
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThresholdProfile"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThresholdProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: Invalid payload
        details:
          type: object
...
    ThresholdProfile:
      type: object
      properties:
        farmId:
          type: string
        plotId:
          type: string
        humidityMin:
          type: number
        humidityMax:
          type: number
        temperatureMin:
          type: number
        temperatureMax:
          type: number
        phMin:
          type: number
        phMax:
          type: number
        ceMin:
          type: number
        ceMax:
          type: number
        nMin:
          type: number
        nMax:
          type: number
        pMin:
          type: number
        pMax:
          type: number
        kMin:
          type: number
        kMax:
          type: number
...